requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
sqlalchemy
prometheus-client
discord.py
pydantic
google-generativeai